+++ reminders/src/lib.rs
@@ -1,5 +1,4 @@
 mod bitmap;
-mod engine;
 pub mod interface;
 mod memory;
 mod schedule;
